src/generation/02 - Expand Image/requirements/requirements-all.txt
# Core libraries
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# Diffusers and Stable Diffusion
diffusers==0.30.0
transformers==4.44.0
accelerate==0.34.0

# Image processing
pillow==11.3.0

# General
pydantic==2.11.9

# Local packages
-e ../../utilities
-e ../../sdxl